+++ main.py
@@ -10,98 +10,9 @@ from obsidian import ObsidianAPI
 from ai import FlashcardAI
 from anki import AnkiAPI
 from config import ConfigManager, MAX_CARDS, NOTES_TO_SAMPLE, DAYS_OLD, SAMPLING_MODE, CARD_TYPE
+from wizard import setup, ENV_FILE, CONFIG_FILE
 
 console = Console()
-
-# Standard config directory location
-CONFIG_DIR = Path.home() / ".config" / "obsidianki"
-ENV_FILE = CONFIG_DIR / ".env"
-CONFIG_FILE = CONFIG_DIR / "config.json"
-
-def setup():
-    """Interactive setup to configure API keys and preferences"""
-    console.print(Panel(Text("ObsidianKi Setup", style="bold blue"), style="blue"))
-
-    step_num = 1
-
-    # Ensure config directory exists
-    CONFIG_DIR.mkdir(parents=True, exist_ok=True)
-
-    # Setup API keys only if .env doesn't exist
-    if not ENV_FILE.exists():
-        console.print(f"[cyan]Step {step_num}: API Keys[/cyan]")
-        console.print("   Get Obsidian API key from: Obsidian Settings > Community Plugins > REST API > API Key")
-        obsidian_key = input("   Enter your Obsidian API key: ").strip()
-
-        console.print("\n   Get Anthropic API key from: https://console.anthropic.com/")
-        anthropic_key = input("   Enter your Anthropic API key: ").strip()
-
-        # Create .env file
-        env_content = f"""OBSIDIAN_API_KEY={obsidian_key}
-        ANTHROPIC_API_KEY={anthropic_key}
-        """
-
-        try:
-            with open(ENV_FILE, "w") as f:
-                f.write(env_content)
-            console.print("   [green]✓[/green] API keys saved")
-        except Exception as e:
-            console.print(f"   [red]ERROR:[/red] Could not create .env file: {e}")
-            return
-        step_num += 1
-    else:
-        console.print("[green]✓[/green] API keys already configured")
-
-    if not CONFIG_FILE.exists():
-        console.print(f"\n[cyan]Step {step_num}: Preferences[/cyan]")
-
-        try:
-            max_cards = int(input("   How many flashcards per session? (default: 6): ").strip() or "6")
-        except ValueError:
-            max_cards = 6
-
-        try:
-            notes_to_sample = int(input("   How many notes to sample? (default: 3): ").strip() or "3")
-        except ValueError:
-            notes_to_sample = 3
-
-        try:
-            days_old = int(input("   Only process notes older than X days? (default: 7): ").strip() or "7")
-        except ValueError:
-            days_old = 7
-
-        sampling_mode = input("   Sampling mode - 'random' or 'weighted'? (default: random): ").strip().lower()
-        if sampling_mode not in ['random', 'weighted']:
-            sampling_mode = 'random'
-
-        card_type = input("   Card type - 'basic' or 'cloze'? (default: basic): ").strip().lower()
-        if card_type not in ['basic', 'cloze']:
-            card_type = 'basic'
-
-        # Create config.json with user preferences
-        user_config = {
-            "MAX_CARDS": max_cards,
-            "NOTES_TO_SAMPLE": notes_to_sample,
-            "DAYS_OLD": days_old,
-            "SAMPLING_MODE": sampling_mode,
-            "CARD_TYPE": card_type,
-            "SEARCH_FOLDERS": []
-        }
-
-        try:
-            import json
-            with open(CONFIG_FILE, "w") as f:
-                json.dump(user_config, f, indent=2)
-            console.print("   [green]✓[/green] Configuration saved")
-        except Exception as e:
-            console.print(f"   [red]ERROR:[/red] Could not create config.json: {e}")
-            return
-    else:
-        console.print("[green]✓[/green] Configuration already exists")
-
-    console.print("\n[green]Setup complete![/green]")
-    console.print(f"[cyan]Config location:[/cyan] {CONFIG_DIR}")
-    console.print("\nYou can now run 'obsidianki' to generate flashcards, or 'obsidianki --setup' to reconfigure.")
 
 def main():
     parser = argparse.ArgumentParser(description="Generate flashcards from Obsidian notes")
